requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pyyaml>=5.4.0
